Refactoring of code, splitting into submodules
